git add -a

+++ lesson_02/lesson_2_task_3.py
@@ -1,7 +1,9 @@
 import math
 
+
 def square(items):
     return math.ceil(items * items)
 
-num_items = int(input("Введите сторону квадрата: "))
+
+num_items = float(input("Введите сторону квадрата: "))
 print(f"Площадь квадрата: {square(num_items)}")
